Feature/order

+++ src/main/resources/sql-schema.sql
@@ -5,14 +5,14 @@ CREATE SCHEMA IF NOT EXISTS `ims`;
 USE `ims` ;
 
 CREATE TABLE IF NOT EXISTS `ims`.`customers` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `first_name` VARCHAR(64) NOT NULL,
     `surname` VARCHAR(64) NOT NULL,
     PRIMARY KEY (`id`)
 );
 
 CREATE TABLE IF NOT EXISTS `ims`.`items` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `type` VARCHAR(40) NOT NULL,
     `name` VARCHAR(40) NOT NULL,
     `description` VARCHAR(256),
@@ -21,8 +21,8 @@ CREATE TABLE IF NOT EXISTS `ims`.`items` (
 );
 
 CREATE TABLE IF NOT EXISTS `ims`.`orders` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
-    `customer_id` INT(11) NOT NULL,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
+    `customer_id` BIGINT NOT NULL,
     `date_placed` DATE DEFAULT(DATE(CURRENT_TIMESTAMP)),
     PRIMARY KEY (`id`),
     FOREIGN KEY (`customer_id`) REFERENCES `customers`(`id`)
@@ -30,8 +30,8 @@ CREATE TABLE IF NOT EXISTS `ims`.`orders` (
 
 
 CREATE TABLE IF NOT EXISTS `ims`.`orderitems` (
-    `order_id` INT(11) NOT NULL,
-    `item_id` INT(11) NOT NULL,
+    `order_id` BIGINT NOT NULL,
+    `item_id` BIGINT NOT NULL,
     `quantity` INT DEFAULT (1),
     PRIMARY KEY (`order_id`, `item_id`),
     FOREIGN KEY (`order_id`) REFERENCES `orders`(`id`),
